Add files via upload
In this thesis, some problems related to probabilistic estimation using inequalities were discussed. In particular, we discussed the derivations of Azuma's, Khinchins and Talagrand's inequalities. In addition to that, we discuss the problem of the concentration of measure phenomenon. The work is a compilation of publications, has modified proofs of elementary theorems, that were complemented with analytical examples of usage: a bin packing problem, a traveling salesman problem and a Steiner tree problem. Furthermore, the connection between the concentration of measure phenomenon and Lipschitz functions was discussed, along with a comparison of Talagrand inequality with other inequalities that emerge from Azuma's inequality.